requirements.txt
flask
flask-cors
torch
torchvision
timm
numpy
pillow